Shorten SED, Bandpass, LookupTable reprs even more. (#723-better-repr)
