+++ galsim/sed.py
@@ -589,10 +589,12 @@ class SED(object):
             wave_type = ' wave_type="Angstroms",'
         if self.flux_type != 'flambda':
             flux_type = ' flux_type=%r,' % self.flux_type
-        return ('galsim.SED(%r, redshift=%r,%s%s' +
-                ' _wave_list=%r, _blue_limit=%r, _red_limit=%r)') % (
-                    self._orig_spec, self.redshift, wave_type, flux_type,
-                    self.wave_list, self.blue_limit, self.red_limit)
+        with galsim.utilities.printoptions(threshold=6):
+            outstr = ('galsim.SED(%r, redshift=%r,%s%s' +
+                      ' _wave_list=%r, _blue_limit=%r, _red_limit=%r)') % (
+                          self._orig_spec, self.redshift, wave_type, flux_type,
+                          self.wave_list, self.blue_limit, self.red_limit)
+        return outstr
 
     def __str__(self):
         orig_spec = repr(self._orig_spec)
